mcs48: the A11 bank bit did not fit in the register holding it
SEL MB1 sets A11 to reach the upper half of a 4 KB program, and a11 was a
uint8_t - so 0x800 truncated to zero and the instruction silently did nothing.
Donkey Kong's sound ROM is 2 KB mirrored, so it never noticed; Gyruss's 8039
has a full 4 KB and does.

+++ core/mcs48.h
@@ -36,7 +36,7 @@ typedef struct {
     uint8_t timer_flag, f1;
     uint8_t irq_enabled, tcnti_enabled;
     uint8_t irq_line, in_irq;
-    uint8_t a11;                 /* SEL MB0 / MB1 */
+    uint16_t a11;                /* SEL MB0 / MB1: the A11 bank bit, so this must hold 0x800 */
     int32_t icount;
 } mcs48_t;
